+++ corvid/table_extraction/pdf_parser.py
@@ -4,7 +4,6 @@ Classes that calls local PDF -> XML parser via subprocess
 
 """
 
-
 import os
 import subprocess
 
@@ -12,8 +11,10 @@ import subprocess
 class PDFParserException(Exception):
     pass
 
+
 class TetmlPDFParserException(PDFParserException):
     pass
+
 
 class OmnipagePDFParserException(PDFParserException):
     pass
@@ -22,13 +23,13 @@ class OmnipagePDFParserException(PDFParserException):
 class PDFParser(object):
     def __init__(self, target_dir: str):
         if not os.path.exists(target_dir):
-            raise FileNotFoundError('Target directory {} doesnt exist'.format(target_dir))
+            raise FileNotFoundError('Target directory {} doesnt exist'
+                                    .format(target_dir))
         self.target_dir = target_dir
 
     def parse(self, paper_id: str, source_pdf_path: str) -> str:
         """Primary method for parsing a Paper's PDF given its id.
         Returns the local path of the XML output of the parsing.
-
         Raises exception unless user implements `_parse()`
         """
         target_path = self.get_target_path(paper_id)
@@ -36,18 +37,23 @@ class PDFParser(object):
         return target_path
 
     def _parse(self, paper_id: str, source_pdf_path: str, target_path: str):
+        """User should implement this.  Recommended to catch Exceptions
+        and instead raise custom Exceptions based on PDFParserException."""
         raise NotImplementedError
 
     def get_target_path(self, paper_id: str) -> str:
+        """Each PDF parser is responsible for constructing output filenames
+        given the paper_id and self.target_dir.  This includes any special
+        nesting directories (e.g. multiple files output given paper_id)"""
         raise NotImplementedError
 
 
 class TetmlPDFParser(PDFParser):
     def __init__(self, tet_bin_path: str, target_dir: str):
         if not os.path.exists(tet_bin_path):
-            raise TetmlPDFParserException('{} doesnt exist'.format(tet_bin_path))
+            raise TetmlPDFParserException(
+                '{} doesnt exist'.format(tet_bin_path))
         self.tet_bin_path = tet_bin_path
-
         super(TetmlPDFParser, self).__init__(target_dir)
 
     def _parse(self, paper_id: str, source_pdf_path: str, target_path: str):
@@ -62,7 +68,8 @@ class TetmlPDFParser(PDFParser):
             if os.path.exists(target_path):
                 os.remove(target_path)
             print(e)
-            raise TetmlPDFParserException('TET failed to parse {}'.format(source_pdf_path))
+            raise TetmlPDFParserException('TET failed to parse {}'
+                                          .format(source_pdf_path))
 
     def get_target_path(self, paper_id: str) -> str:
         return '{}.tetml'.format(os.path.join(self.target_dir, paper_id))
@@ -71,9 +78,9 @@ class TetmlPDFParser(PDFParser):
 class OmnipagePDFParser(PDFParser):
     def __init__(self, omnipage_bin_path: str, target_dir: str):
         if not os.path.exists(omnipage_bin_path):
-            raise OmnipagePDFParserException('{} doesnt exist'.format(omnipage_bin_path))
+            raise OmnipagePDFParserException('{} doesnt exist'
+                                             .format(omnipage_bin_path))
         self.omnipage_bin_path = omnipage_bin_path
-
         super(OmnipagePDFParser, self).__init__(target_dir)
 
     def _parse(self, paper_id: str, source_pdf_path: str, target_path: str):
@@ -85,8 +92,8 @@ class OmnipagePDFParser(PDFParser):
             subprocess.run(cmd, shell=True, check=True)
         except subprocess.CalledProcessError as e:
             print(e)
-            raise OmnipagePDFParserException('OmniPage failed to parse {}'.format(source_pdf_path))
+            raise OmnipagePDFParserException('OmniPage failed to parse {}'
+                                             .format(source_pdf_path))
 
     def get_target_path(self, paper_id: str) -> str:
         return '{}.xml'.format(os.path.join(self.target_dir, paper_id))
-
